backend/requirements.txt
flask-cors==6.0.0
numpy>=1.26.2
pandas>=2.1.1
python-dotenv==1.0.0
requests>=2.31.0
schedule==1.2.2
scikit-learn>=1.3.1
tensorflow>=2.15.0
PyJWT>=2.8.0
werkzeug>=2.3.0